меняет отрисовку комментариев
