--- src/view/comment-view.js
@@ -1,32 +1,0 @@
-import AbstractView from './abstract-view';
-
-const createCommentTemplate = (comment) => {
-  const { emotion, commentText, name, date } = comment;
-
-  return `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="${emotion}" width="55" height="55" alt="emoji-smile">
-    </span>
-    <div>
-      <p class="film-details__comment-text">${commentText}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${name}</span>
-        <span class="film-details__comment-day">${date}</span>
-        <button class="film-details__comment-delete">Delete</button>
-      </p>
-    </div>
-  </li>`;
-};
-
-export default class CommentView extends AbstractView {
-  #comment = null;
-
-  constructor(comment) {
-    super();
-    this.#comment = comment;
-  }
-
-  get template() {
-    return createCommentTemplate(this.#comment);
-  }
-}
