+++ src/view/films-popup-view.js
@@ -9,9 +9,33 @@ const createGenresTemplate = (genres) => {
   return list;
 };
 
+const createCommentTemplate = (comments) => {
+  let list = '';
+  for (const comment of comments) {
+    const { emotion, commentText, name, date } = comment;
+
+    list += `<li class="film-details__comment">
+            <span class="film-details__comment-emoji">
+              <img src="${emotion}" width="55" height="55" alt="emoji-smile">
+            </span>
+            <div>
+              <p class="film-details__comment-text">${commentText}</p>
+              <p class="film-details__comment-info">
+                <span class="film-details__comment-author">${name}</span>
+                <span class="film-details__comment-day">${date}</span>
+                <button class="film-details__comment-delete">Delete</button>
+              </p>
+            </div>
+          </li>`;
+  }
+  return list;
+};
+
 const createFilmsPopupTemplate = (card) => {
-  const { poster, title, rating, duration, genres, age, director, writers, actors, dateRelease, isAddedToWatch, isWatched, isFavorite } = card;
+  const { poster, comments, title, rating, duration, genres, age, director, writers, actors, dateRelease, isAddedToWatch, isWatched, isFavorite } = card;
   const itemsGenres = createGenresTemplate(genres);
+  const itemsComments = createCommentTemplate(comments);
+  const count = comments.length;
 
   const addWatchListClassName = isAddedToWatch
     ? 'film-details__control-button--active'
@@ -97,10 +121,10 @@ const createFilmsPopupTemplate = (card) => {
 
     <div class="film-details__bottom-container">
       <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">4</span></h3>
+        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${count}</span></h3>
 
         <ul class="film-details__comments-list">
-
+        ${itemsComments}
         </ul>
 
         <div class="film-details__new-comment">
